Refactor the publishing process
This is prompted by pull request #16, and incorporates many of the ideas there. I decided TwineApp was the right place to put these methods since they are used in both StoryListView and StoryEditView, and now all publishing is routed through its publishStory() method. I also made saveFile() a separate method so that if we want to save other kinds of files in the future (I'm thinking of a storymap export), we have more flexibility there.