Story format passage syntax highlighting support.
This extends the format.js format, adding a "lexer" property which consists of a stringified function. This function is capable of defining a CodeMirror mode for passages in that story format. (However, it's also intended for providing future features related to lexing, like early error diagnostics, or special passage arrow/tag colouring.)

This also extends PassageEditView to support using that CodeMirror mode for the passage editor pane if it exists, and defaulting to text if it does not.